Reorganize the structure of `generateBefore`
