+++ scripts/events/lib/injects-point.js
