+++ scripts/events/lib/injects.js
